+++ src/models/scallopCache.ts
@@ -1,4 +1,4 @@
-import { QueryClient, QueryClientConfig } from '@tanstack/query-core';
+import { QueryClient } from '@tanstack/query-core';
 import {
   SuiObjectArg,
   SuiTxBlock,
@@ -17,19 +17,43 @@ import type {
   GetDynamicFieldObjectParams,
   GetBalanceParams,
   SuiClient,
+  CoinBalance,
 } from '@mysten/sui/client';
 import { DEFAULT_CACHE_OPTIONS } from 'src/constants/cache';
-import { callWithRateLimit, TokenBucket } from 'src/utils';
-import {
-  DEFAULT_INTERVAL_IN_MS,
-  DEFAULT_TOKENS_PER_INTERVAL,
-} from 'src/constants/tokenBucket';
 import { queryKeys } from 'src/constants';
+import { ScallopCacheInstanceParams, ScallopCacheParams } from 'src/types';
+import { newSuiKit } from './suiKit';
 
 type QueryInspectTxnParams = {
   queryTarget: string;
   args: SuiObjectArg[];
   typeArgs?: any[];
+};
+
+const DEFAULT_TOKENS_PER_INTERVAL = 10;
+const DEFAULT_INTERVAL_IN_MS = 250;
+
+const deepMergeObject = <T>(curr: T, update: T): T => {
+  const result = { ...curr }; // Clone the current object to avoid mutation
+
+  for (const key in update) {
+    if (
+      update[key] &&
+      typeof update[key] === 'object' &&
+      !Array.isArray(update[key])
+    ) {
+      // If the value is an object, recurse
+      result[key] = deepMergeObject(
+        curr[key] || ({} as T[Extract<keyof T, string>]),
+        update[key]
+      );
+    } else {
+      // Otherwise, directly assign the value
+      result[key] = update[key];
+    }
+  }
+
+  return result;
 };
 
 /**
@@ -46,20 +70,26 @@ type QueryInspectTxnParams = {
  */
 
 export class ScallopCache {
-  public readonly queryClient: QueryClient;
-  public readonly _suiKit: SuiKit;
-  private tokenBucket: TokenBucket;
+  public readonly params: ScallopCacheParams;
+
+  public queryClient: QueryClient;
+  public suiKit: SuiKit;
+  // private tokenBucket: TokenBucket;
   public walletAddress: string;
+  private tokensPerInterval: number = DEFAULT_TOKENS_PER_INTERVAL;
+  private interval: number = DEFAULT_INTERVAL_IN_MS;
+  private tokens: number;
+  private lastRefill: number;
 
   public constructor(
-    suiKit: SuiKit,
-    walletAddress?: string,
-    cacheOptions?: QueryClientConfig,
-    tokenBucket?: TokenBucket,
-    queryClient?: QueryClient
+    params: ScallopCacheParams = {},
+    instance?: ScallopCacheInstanceParams
   ) {
+    this.params = params;
+    this.suiKit = instance?.suiKit ?? newSuiKit(params);
     this.queryClient =
-      queryClient ?? new QueryClient(cacheOptions ?? DEFAULT_CACHE_OPTIONS);
+      instance?.queryClient ??
+      new QueryClient(params?.cacheOptions ?? DEFAULT_CACHE_OPTIONS);
 
     // handle case where there's existing queryClient and cacheOptions is also passed
     // if (queryClient && cacheOptions) {
@@ -71,22 +101,62 @@ export class ScallopCache {
     //   // if(cacheOptions.mutations)this.queryClient.setMutationDefaults(cacheOptions.mutations);
     // }
 
-    this._suiKit = suiKit;
-    this.tokenBucket =
-      tokenBucket ??
-      new TokenBucket(DEFAULT_TOKENS_PER_INTERVAL, DEFAULT_INTERVAL_IN_MS);
-    this.walletAddress = walletAddress ?? suiKit.currentAddress();
-  }
-
-  private get suiKit(): SuiKit {
-    if (!this._suiKit) {
-      throw new Error('SuiKit instance is not initialized');
-    }
-    return this._suiKit;
+    this.tokens = this.tokensPerInterval;
+    this.lastRefill = Date.now();
+    this.walletAddress = params.walletAddress ?? this.suiKit.currentAddress();
   }
 
   private get client(): SuiClient {
     return this.suiKit.client();
+  }
+
+  private refill() {
+    const now = Date.now();
+    const elapsed = now - this.lastRefill;
+
+    if (elapsed >= this.interval) {
+      const tokensToAdd =
+        Math.floor(elapsed / this.interval) * this.tokensPerInterval;
+      this.tokens = Math.min(this.tokens + tokensToAdd, this.tokensPerInterval);
+
+      // Update lastRefill to reflect the exact time of the last "refill"
+      this.lastRefill += Math.floor(elapsed / this.interval) * this.interval;
+    }
+  }
+
+  private removeTokens(count: number) {
+    this.refill();
+    if (this.tokens >= count) {
+      this.tokens -= count;
+      return true;
+    }
+    return false;
+  }
+
+  private async callWithRateLimit<T>(
+    fn: () => Promise<T>,
+    maxRetries = 15,
+    backoffFactor = 1.25 // The factor by which to increase the delay
+  ): Promise<T | null> {
+    let retries = 0;
+
+    const tryRequest = async (): Promise<T | null> => {
+      if (this.removeTokens(1)) {
+        const result = await fn();
+        return result;
+      } else if (retries < maxRetries) {
+        retries++;
+        const delay = this.interval * Math.pow(backoffFactor, retries);
+        // console.error(`Rate limit exceeded, retrying in ${delay} ms`);
+        await new Promise((resolve) => setTimeout(resolve, delay));
+        return tryRequest();
+      } else {
+        console.error('Maximum retries reached');
+        return null;
+      }
+    };
+
+    return tryRequest();
   }
 
   /**
@@ -168,8 +238,7 @@ export class ScallopCache {
       retryDelay: 1000,
       queryKey: queryKeys.rpc.getInspectTxn(queryTarget, args, typeArgs),
       queryFn: async () => {
-        return await callWithRateLimit(
-          this.tokenBucket,
+        return await this.callWithRateLimit(
           async () => await this.suiKit.inspectTxn(txBlock)
         );
       },
@@ -182,8 +251,7 @@ export class ScallopCache {
     return this.queryClient.fetchQuery({
       queryKey: queryKeys.rpc.getNormalizedMoveFunction(target),
       queryFn: async () => {
-        return await callWithRateLimit(
-          this.tokenBucket,
+        return await this.callWithRateLimit(
           async () =>
             await this.suiKit.client().getNormalizedMoveFunction({
               package: address,
@@ -209,14 +277,14 @@ export class ScallopCache {
       ...options,
       showOwner: true,
       showContent: true,
+      showType: true,
     };
     return this.queryClient.fetchQuery({
       retry: this.retryFn,
       retryDelay: 1000,
       queryKey: queryKeys.rpc.getObject(objectId, options),
       queryFn: async () => {
-        return await callWithRateLimit(
-          this.tokenBucket,
+        return await this.callWithRateLimit(
           async () =>
             await this.client.getObject({
               id: objectId,
@@ -232,43 +300,38 @@ export class ScallopCache {
    * @param objectIds
    * @returns Promise<SuiObjectData[]>
    */
-  public async queryGetObjects(
-    objectIds: string[],
-    options: SuiObjectDataOptions = {
-      showContent: true,
-    }
-  ): Promise<SuiObjectData[]> {
+  public async queryGetObjects(objectIds: string[]): Promise<SuiObjectData[]> {
     if (objectIds.length === 0) return [];
-    // objectIds.sort();
+    const options: SuiObjectDataOptions = {
+      showContent: true,
+      showOwner: true,
+      showType: true,
+    };
 
     return this.queryClient.fetchQuery({
       retry: this.retryFn,
       retryDelay: 1000,
-      queryKey: queryKeys.rpc.getObjects(
-        objectIds,
-        this.walletAddress,
-        options
-      ),
+      queryKey: queryKeys.rpc.getObjects(objectIds),
       queryFn: async () => {
-        const results = await callWithRateLimit(
-          this.tokenBucket,
+        const results = await this.callWithRateLimit(
           async () => await this.suiKit.getObjects(objectIds, options)
         );
         if (results) {
           results.forEach((result) => {
-            this.queryClient.setQueriesData(
-              {
+            // fetch previous data
+            const queryKey = queryKeys.rpc.getObject(result.objectId);
+            const prevDatas =
+              this.queryClient.getQueriesData<SuiObjectResponse>({
                 exact: false,
-                queryKey: queryKeys.rpc.getObject(result.objectId, options),
-              },
-              {
-                data: result,
-                error: null,
-              },
-              {
-                updatedAt: Date.now(),
-              }
-            );
+                queryKey,
+              });
+            prevDatas.forEach(([key, prevData]) => {
+              this.queryClient.setQueryData(
+                key,
+                deepMergeObject(prevData, { data: result, error: null }),
+                { updatedAt: Date.now() }
+              );
+            });
           });
         }
         return results;
@@ -282,14 +345,13 @@ export class ScallopCache {
    * @returns Promise<PaginatedObjectsResponse>
    */
   public async queryGetOwnedObjects(input: GetOwnedObjectsParams) {
-    // TODO: This query need its own separate rate limiter (as owned objects can theoretically be infinite), need a better way to handle this
+    // @TODO: This query need its own separate rate limiter (as owned objects can theoretically be infinite), need a better way to handle this
     return this.queryClient.fetchQuery({
       retry: this.retryFn,
       retryDelay: 1000,
       queryKey: queryKeys.rpc.getOwnedObjects(input),
       queryFn: async () => {
-        const results = await callWithRateLimit(
-          this.tokenBucket,
+        const results = await this.callWithRateLimit(
           async () => await this.client.getOwnedObjects(input)
         );
         if (results && results.data.length > 0) {
@@ -301,22 +363,20 @@ export class ScallopCache {
                 NonNullable<{ data: SuiObjectData }> => !!result.data
             )
             .forEach((result) => {
-              this.queryClient.setQueriesData(
-                {
+              // fetch previous data
+              const queryKey = queryKeys.rpc.getObject(result.data.objectId);
+              const prevDatas =
+                this.queryClient.getQueriesData<SuiObjectResponse>({
                   exact: false,
-                  queryKey: queryKeys.rpc.getObject(
-                    result.data.objectId,
-                    input.options ?? {}
-                  ),
-                },
-                {
-                  data: result.data,
-                  error: null,
-                },
-                {
-                  updatedAt: Date.now(),
-                }
-              );
+                  queryKey,
+                });
+              prevDatas.forEach(([key, prevData]) => {
+                this.queryClient.setQueryData(
+                  key,
+                  deepMergeObject(prevData, { data: result.data, error: null }),
+                  { updatedAt: Date.now() }
+                );
+              });
             });
         }
         return results;
@@ -332,8 +392,7 @@ export class ScallopCache {
       retryDelay: 1000,
       queryKey: queryKeys.rpc.getDynamicFields(input),
       queryFn: async () => {
-        return await callWithRateLimit(
-          this.tokenBucket,
+        return await this.callWithRateLimit(
           async () => await this.client.getDynamicFields(input)
         );
       },
@@ -348,26 +407,22 @@ export class ScallopCache {
       retryDelay: (attemptIndex) => Math.min(1000 * attemptIndex, 8000),
       queryKey: queryKeys.rpc.getDynamicFieldObject(input),
       queryFn: async () => {
-        const result = await callWithRateLimit(this.tokenBucket, () =>
+        const result = await this.callWithRateLimit(() =>
           this.client.getDynamicFieldObject(input)
         );
         if (result?.data) {
-          this.queryClient.setQueriesData(
-            {
-              exact: false,
-              queryKey: queryKeys.rpc.getObject(result?.data.objectId, {
-                showContent: true,
-                showOwner: true,
-              }),
-            },
-            {
-              data: result.data,
-              error: null,
-            },
-            {
-              updatedAt: Date.now(),
-            }
-          );
+          const queryKey = queryKeys.rpc.getObject(result.data.objectId);
+          const prevDatas = this.queryClient.getQueriesData<SuiObjectResponse>({
+            exact: false,
+            queryKey,
+          });
+          prevDatas.forEach(([key, prevData]) => {
+            this.queryClient.setQueryData(
+              key,
+              deepMergeObject(prevData, { data: result.data, error: null }),
+              { updatedAt: Date.now() }
+            );
+          });
         }
         return result;
       },
@@ -376,26 +431,24 @@ export class ScallopCache {
 
   public async queryGetAllCoinBalances(
     owner: string
-  ): Promise<{ [k: string]: string }> {
+  ): Promise<{ [k: string]: CoinBalance }> {
     return this.queryClient.fetchQuery({
       retry: this.retryFn,
       retryDelay: 1000,
       queryKey: queryKeys.rpc.getAllCoinBalances(owner),
       queryFn: async () => {
-        const allBalances = await callWithRateLimit(
-          this.tokenBucket,
+        const allBalances = await this.callWithRateLimit(
           async () => await this.client.getAllBalances({ owner })
         );
         if (!allBalances) return {};
         const balances = allBalances.reduce(
           (acc, coinBalance) => {
             if (coinBalance.totalBalance !== '0') {
-              acc[normalizeStructTag(coinBalance.coinType)] =
-                coinBalance.totalBalance;
+              acc[normalizeStructTag(coinBalance.coinType)] = coinBalance;
             }
             return acc;
           },
-          {} as { [k: string]: string }
+          {} as { [k: string]: CoinBalance }
         );
 
         return balances;
@@ -403,8 +456,10 @@ export class ScallopCache {
     });
   }
 
-  public async queryGetCoinBalance(input: GetBalanceParams): Promise<string> {
-    if (!input.coinType) return '0';
+  public async queryGetCoinBalance(
+    input: GetBalanceParams
+  ): Promise<CoinBalance | null> {
+    if (!input.coinType) return null;
 
     return (
       ((await this.queryGetAllCoinBalances(input.owner)) ?? {})[